pyproject.toml
[project]
name = "NameMatcher"
version = "0.1.0"
description = "A robust name matching system that leverages Google Generative AI embeddings and ChromaDB for efficient similarity search and ranking of person names."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
